+++ lib/vagrant/action/box/download.rb
@@ -26,13 +26,25 @@ module Vagrant
         end
 
         def instantiate_downloader
-          @env["download.classes"].each do |klass|
-            if klass.match?(@env["box"].uri)
+          # Assign to a temporary variable since this is easier to type out,
+          # since it is used so many times.
+          classes = @env["download.classes"]
+
+          # Find the class to use.
+          classes.each_index do |i|
+            klass = classes[i]
+
+            # Use the class if it matches the given URI or if this
+            # is the last class...
+            if classes.length == (i + 1) || klass.match?(@env["box"].uri)
               @env.ui.info I18n.t("vagrant.actions.box.download.with", :class => klass.to_s)
               @downloader = klass.new(@env)
+              break
             end
           end
 
+          # This line should never be reached, but we'll keep this here
+          # just in case for now.
           raise Errors::BoxDownloadUnknownType if !@downloader
 
           @downloader.prepare(@env["box"].uri)