+++ lib/vagrant/util.rb
@@ -2,7 +2,6 @@ module Vagrant
   module Util
     autoload :Busy,                      'vagrant/util/busy'
     autoload :Counter,                   'vagrant/util/counter'
-    autoload :GlobLoader,                'vagrant/util/glob_loader'
     autoload :HashWithIndifferentAccess, 'vagrant/util/hash_with_indifferent_access'
     autoload :Platform,                  'vagrant/util/platform'
     autoload :Retryable,                 'vagrant/util/retryable'