+++ test/buildbot/buildbot_config/master/change_sources.py
@@ -1,0 +1,8 @@
+"""
+Contains the logic to build and return the change sources for
+the build master.
+"""
+
+def get_change_sources():
+    # Currently no sources since we're using a GitHub hook
+    return []