+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 0.8.8 (unreleased)
 
-
+  - Mount shared folders shortest to longest to avoid mounting
+    subfolders first. [GH-525]
+  - Support for basic HTTP auth in the URL for boxes.
+  - Solaris support for host only networks. [GH-533]
+  - `vagrant init` respects `Vagrant::Environment` cwd. [GH-528]
+  - `vagrant` commands will not output color when stdout is
+    not a TTY.
 
 ## 0.8.7 (September 13, 2011)